auth problems fixed
