+++ app/build.gradle
@@ -20,6 +20,8 @@ android {
 }
 
 dependencies {
+    implementation 'com.google.firebase:firebase-auth:17.0.0'
+    implementation 'com.firebaseui:firebase-ui-database:6.0.2'
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation 'androidx.appcompat:appcompat:1.0.0-beta01'
     implementation 'com.google.firebase:firebase-auth:19.2.0'
