+++ app/src/main/java/com/example/myapplication/MainActivity.java
@@ -1,11 +1,21 @@
 package com.example.myapplication;
 
+import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
+
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
+import android.widget.Toast;
+
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.FirebaseApp;
+import com.google.firebase.auth.AuthResult;
+import com.google.firebase.auth.FirebaseAuth;
 
 import Objects.Market;
 
@@ -37,14 +47,42 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onClick(View v) {
                 //put stuff for player auth
-                Intent intent = new Intent(context, Sign_up_player.class);
-                context.startActivity(intent);
+                FirebaseAuth auth = FirebaseAuth.getInstance();
+
+
+                auth.signInAnonymously().addOnCompleteListener(new OnCompleteListener<AuthResult>() {
+                    @Override
+                    public void onComplete(@NonNull Task<AuthResult> task) {
+                        if (task.isSuccessful()){
+
+                            String player = FirebaseAuth.getInstance().getCurrentUser().getUid();
+                            if (player!=null && FirebaseAuth.getInstance().getCurrentUser().isAnonymous()){
+                            Intent intent = new Intent(context, Wait_Page.class);
+                            intent.putExtra("player_id", player);
+                            context.startActivity(intent);
+                            finish();}
+                            else{
+                                Toast.makeText(context, "Error in signing in", Toast.LENGTH_SHORT).show();
+                            };
+                        }
+                        else{
+                            Toast.makeText(context, "Error in signing in", Toast.LENGTH_SHORT).show();
+                        };
+
+                    }
+                });
+
+
+
+
+
 
             }
         });
 
-
     }
+
+
 }
