+++ app/src/main/java/com/example/myapplication/Set_Parameters_Round_2.java
@@ -1,23 +1,21 @@
 package com.example.myapplication;
 
-import androidx.appcompat.app.AppCompatActivity;
-
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
 
 import com.google.firebase.auth.FirebaseAuth;
 
-public class Set_Parameters extends AppCompatActivity {
-    FirebaseAuth mauth;
+import androidx.appcompat.app.AppCompatActivity;
+
+public class Set_Parameters_Round_2 extends AppCompatActivity {
     Context context;
-
-
+    FirebaseAuth mauth;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_set__parameters);
+        setContentView(R.layout.activity_set__parameters_round_2);
         context=(Context) this;
         mauth = FirebaseAuth.getInstance();
         if (mauth.getCurrentUser().isAnonymous()|| mauth.getCurrentUser()==null){
