+++ app/src/main/res/layout/activity_set__parameters_practice.xml
@@ -5,9 +5,9 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".Set_Parameters">
-    <LinearLayout
 
-        android:id="@+id/set_round_1"
+    <LinearLayout
+        android:id="@+id/set_practice_round"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent"
@@ -23,7 +23,7 @@
             android:textAllCaps="true"
             android:textColor="@color/colorPrimaryDark"
             android:textSize= "@dimen/textSizeTitle"
-            android:text="ROUND 1"
+            android:text="PRACTICE ROUND "
             />
 
 
@@ -38,7 +38,7 @@
             android:text="Enter p"
             />
         <EditText
-            android:id="@+id/prob_ofhigh"
+            android:id="@+id/prob_ofhigh_practice"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:gravity="right"
@@ -57,7 +57,7 @@
             android:text="Enter pi_h"
             />
         <EditText
-            android:id="@+id/prob_outcome_high"
+            android:id="@+id/prob_outcome_high_practice"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:gravity="right"
@@ -73,10 +73,10 @@
             android:textAllCaps="true"
             android:textColor="@color/colorPrimaryDark"
             android:textSize= "@dimen/textSizeNormal"
-            android:text="Enter pi_l"
+            android:text="pi_L"
             />
         <EditText
-            android:id="@+id/prob_outcome_low"
+            android:id="@+id/prob_outcome_low_practice"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:gravity="right"
@@ -92,10 +92,10 @@
             android:textAllCaps="true"
             android:textColor="@color/colorPrimaryDark"
             android:textSize= "@dimen/textSizeNormal"
-            android:text="Enter number of rounds"
+            android:text="Enter number  of Rounds"
             />
         <EditText
-            android:id="@+id/num_rounds"
+            android:id="@+id/num_rounds_practice"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:gravity="right"
@@ -104,9 +104,11 @@
             android:textColor="@color/colorPrimaryDark"
             android:textSize= "@dimen/textSizeNormal"/>
 
+
     </LinearLayout>
+
     <Button
-        android:id="@+id/button_submit_round_1"
+        android:id="@+id/button_submit_practice_round"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toBottomOf="@id/set_practice_round"
